🎨 fetchMyProfile 분리
